Change select id, move components to common, fixed remain useIntl function call
