+++ src/components/common/SelectForm.js
@@ -1,7 +1,7 @@
 import React from "react";
 
-const SelectForm = ({ options }) => (
-  <select id="genderSelect" className="form-select">
+const SelectForm = ({ options, id }) => (
+  <select id={id} className="form-select">
     {options.map((option) => (
       <option value={option.value}>{option.text}</option>
     ))}
